% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_registry.R
\name{db_list_sources}
\alias{db_list_sources}
\title{List all registered Python scraper sources}
\usage{
db_list_sources()
}
\value{
Character vector of source names.
}
\description{
List all registered Python scraper sources
}
