% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downballot_install_python.R
\name{db_stop_if_python_initialized_to_other}
\alias{db_stop_if_python_initialized_to_other}
\title{Internal: stop if reticulate initialized to a different python}
\usage{
db_stop_if_python_initialized_to_other(envname = "downballotR")
}
\description{
Internal: stop if reticulate initialized to a different python
}
\keyword{internal}
