% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_registry.R
\name{.db_registry}
\alias{.db_registry}
\title{Import the Python registry module (after ensuring Python is ready)}
\usage{
.db_registry()
}
\description{
Import the Python registry module (after ensuring Python is ready)
}
\keyword{internal}
