% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_registry.R
\name{.scrape_ballotpedia}
\alias{.scrape_ballotpedia}
\title{Call the Ballotpedia school board scraper}
\usage{
.scrape_ballotpedia(
  year = NULL,
  state = NULL,
  mode = "districts",
  start_year = 2013L,
  end_year = NULL
)
}
\description{
Call the Ballotpedia school board scraper
}
\keyword{internal}
