% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_registry.R
\name{.scrape_ct}
\alias{.scrape_ct}
\title{Call the Connecticut CTEMS election results scraper}
\usage{
.scrape_ct(
  year_from = NULL,
  year_to = NULL,
  level = "all",
  max_town_workers = 2L
)
}
\description{
Call the Connecticut CTEMS election results scraper
}
\keyword{internal}
