% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_registry.R
\name{.scrape_election_stats}
\alias{.scrape_election_stats}
\title{Call the ElectionStats scraper}
\usage{
.scrape_election_stats(
  state,
  year_from = 1789L,
  year_to = NULL,
  level = "all",
  parallel = TRUE
)
}
\description{
Call the ElectionStats scraper
}
\keyword{internal}
