% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_registry.R
\name{.scrape_ga}
\alias{.scrape_ga}
\title{Call the Georgia SOS election results scraper}
\usage{
.scrape_ga(
  year_from = NULL,
  year_to = NULL,
  level = "all",
  max_county_workers = 4L,
  include_vote_methods = FALSE
)
}
\description{
Call the Georgia SOS election results scraper
}
\keyword{internal}
