% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_helpers.R
\name{.scrape_in}
\alias{.scrape_in}
\title{Call the Indiana General Election results scraper}
\usage{
.scrape_in(year_from = NULL, year_to = NULL, level = "all")
}
\description{
Call the Indiana General Election results scraper
}
\keyword{internal}
