% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_helpers.R
\name{.scrape_la}
\alias{.scrape_la}
\title{Call the Louisiana Secretary of State election results scraper}
\usage{
.scrape_la(
  year_from = NULL,
  year_to = NULL,
  level = "all",
  max_parish_workers = 2L
)
}
\description{
Call the Louisiana Secretary of State election results scraper
}
\keyword{internal}
