% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_helpers.R
\name{.scrape_ut}
\alias{.scrape_ut}
\title{Call the Utah election results scraper}
\usage{
.scrape_ut(
  year_from = NULL,
  year_to = NULL,
  level = "all",
  max_county_workers = 4L,
  include_vote_methods = FALSE
)
}
\description{
Call the Utah election results scraper
}
\keyword{internal}
