% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_utils.R
\name{.stop_if_not_scalar}
\alias{.stop_if_not_scalar}
\title{Stop if a scalar argument has length != 1}
\usage{
.stop_if_not_scalar(x, arg)
}
\description{
Stop if a scalar argument has length != 1
}
\keyword{internal}
