% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downballot_use_python.R
\name{downballot_use_python}
\alias{downballot_use_python}
\title{Use the downballotR Python virtualenv in this R session}
\usage{
downballot_use_python(envname = "downballotR")
}
\arguments{
\item{envname}{Name of the virtualenv to use.}
}
\value{
Invisibly TRUE on success.
}
\description{
Pins reticulate to the package's virtualenv for the current R session.
If reticulate is already initialized to a different interpreter, this will
error with a clear message (because reticulate cannot switch interpreters
mid-session).
}
