% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_snapshot.R
\name{nc_missing_dates}
\alias{nc_missing_dates}
\title{Check which NC election dates are missing from the snapshot}
\usage{
nc_missing_dates(dates)
}
\arguments{
\item{dates}{vector of Date (election days)}
}
\value{
list(covered=..., missing=...)
}
\description{
Check which NC election dates are missing from the snapshot
}
