% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_snapshot.R
\name{nc_snapshot_dates}
\alias{nc_snapshot_dates}
\title{List election dates covered by the shipped snapshot (fast)}
\usage{
nc_snapshot_dates()
}
\value{
a vector of Date
}
\description{
List election dates covered by the shipped snapshot (fast)
}
