% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downballot_python_status.R
\name{print.downballot_python_status}
\alias{print.downballot_python_status}
\title{Print a downballot_python_status object}
\usage{
\method{print}{downballot_python_status}(x, ...)
}
\arguments{
\item{x}{A \code{downballot_python_status} object.}

\item{...}{Further arguments passed to or from other methods (unused).}
}
\value{
Invisibly returns \code{x}, the \code{downballot_python_status}
object passed in, following the S3 print method convention.
}
\description{
Print a downballot_python_status object
}
